backend/requirements.txt
flask
requests
Flask-Cors
twilio
pandas
python-dotenv
flask-sock
gunicorn
gevent-websocket 